Updated to handle changes in website HTML
Previously it was simple to select the IYD names box. Now the selector combination we were using before doesn't match anything, and I don't see a unique combination for the IYD box.
Now the selector picks all the .hmy-content nodes, and then we filter to choose only the one that contains "IYD" in its text content.
We also need to filter out non-<P> tags when parsing the nodes inside, to ignore the H2 they have inside the main div.