Removing references to calendar in readme
They've removed actual dates from their website, making it much more difficult to determine what day they're talking about when things change. Now if the "Monday" line is updated, did they mean that change for the current week or the next? Did they just fix a typo or decide to change the name for the current Monday, or is this an update for the next Monday?

I've decided this is too much hassle to deal with, so I'm going to remove the calendaring portion of this program and just have it email diffs when the text changes - maybe the humans receiving the emails will be able to figure out which day they're talking about.